Update Get-WUGDevices.ps1
removed 
else {
        $SearchValue = Read-Host "Enter the search value either IP, hostname, or display name."
    }

because with that in there, it makes the $SearchValue mandatory.  You have to put something there.  Without it, it can still be added as an optional attribute to the Get-WUGDevices.  Now you can just select a group and go with it, or select a group and do a search ...